Upgrade Relay and consolidate resolvers
